ble/alerts/BluetoothNotAvailable: add browser support paragraph
Browsers such as Brave and Opera are based on Chromium but disable
Web Bluetooth. In this case, we can't tell the difference between the
browser not supporting it or the Bluetooth adapter not being present.
So we make the error message cover all possibilities.

Issue: https://github.com/pybricks/support/discussions/804
Issue: https://github.com/pybricks/support/issues/635